Add the ability to skip dependencies when updating

+++ install/rhel-venv.sh
@@ -24,40 +24,51 @@ srcroot="$(readlink -fn "$(dirname "$0")/..")"
 cd $srcroot
 
 # If forced, remove all the things
-[ "$1" == "--force" ] && rm -rf bin/{activate*,easy_install*,pip*,python*} \
-                                include/ lib/ lib64
+if [ "$1" == "--force" ] || [ "$2" == "--force" ]
+then
+    rm -rf bin/{activate*,easy_install*,pip*,python*} \
+           include/ lib/ lib64
+fi
 
-# Install Python
-sudo yum -y install bash curl git python3
+# Prepare the system
+#   For convenience, these can be skipped when you're running them from your
+#   IDE, so you won't need to stay elevated or keep bashing out your password!
+if [ "$1" != "--skip-deps" ] && [ "$2" != "--skip-deps" ]
+then
+    # Install system packages
+    sudo yum -y install bash curl git python3
+
+    # Set up setuptools/distribute
+    pushd /tmp
+    sudo python3.2 < <(curl -s http://python-distribute.org/distribute_setup.py)
+    rm -f distribute_setup.py
+    popd
+
+    # Set up virtualenv
+    sudo easy_install-3.2 virtualenv
+
+    # Set up elevator
+    pushd deps/elevator
+    info=( $(grep -P "^$USER:" /etc/passwd | sed 's/:/ /g' | awk '{print $3, $4}') )
+    ./configure --prefix=../../chroot \
+                --allow-uid="${info[0]}" \
+                --allow-gid="${info[1]}" \
+                --target-uid=0 \
+                --target-gid=0
+    make
+    sudo make install
+    popd
+fi
 
 # Update working copy and initialise submodules (elevator, gnupg, etc.)
 git pull origin master
 git submodule init
 git submodule update
 
-# Set up setuptools/distribute
-pushd /tmp
-sudo python3.2 < <(curl -s http://python-distribute.org/distribute_setup.py)
-rm -f distribute_setup.py
-popd
-
 # Set up virtualenv
 pushd chroot
-sudo easy_install-3.2 virtualenv
 [ ! -f "bin/activate" ] && virtualenv --no-site-packages .
 . bin/activate
-popd
-
-# Set up elevator
-pushd deps/elevator
-info=( $(grep -P "^$USER:" /etc/passwd | sed 's/:/ /g' | awk '{print $3, $4}') )
-./configure --prefix=../../chroot \
-            --allow-uid="${info[0]}" \
-            --allow-gid="${info[1]}" \
-            --target-uid=0 \
-            --target-gid=0
-make
-sudo make install
 popd
 
 # Set up gnupg
